A few small changes. No public API changes.
Added error handling for url.Parse. This shouldn't really be an
issue given the IngestBaseURL is a const. But if for some reason this
needs to change or is changed for testing, I want to make sure it doesn't
fail at this point.

Added a Makefile to build / install / clean and run tests. Created a build
directory for local builds.